chore(discover): Split up the test_discover file
- This file has grown so large that my editor was lagging trying to edit
  it
- this change only moved the test cases around, no changes were made to
  them
